+++ src/view-extensions/references-cm6.ts
@@ -49,7 +49,11 @@ export const InlineReferenceExtension = ViewPlugin.fromClass(class {
                 const mdView = view.state.field( editorInfoField );
                 const firstCharacterMatch = match[0].charAt(0);
                 const transformedCache = getSNWCacheByFile(mdView.file);  
-                if(transformedCache?.cacheMetaData?.frontmatter?.["snw-file-exclude"]!=true) {
+                console.log("transformedCache?.cacheMetaData?.frontmatter", transformedCache?.cacheMetaData?.frontmatter,
+                transformedCache?.cacheMetaData?.frontmatter?.["snw-canvas-exclude-edit"],
+                transformedCache?.cacheMetaData?.frontmatter?.["snw-canvas-exclude-edit"]!=true)
+                if(transformedCache?.cacheMetaData?.frontmatter?.["snw-file-exclude"]!=true && 
+                   transformedCache?.cacheMetaData?.frontmatter?.["snw-canvas-exclude-edit"]!=true) {
                     const widgetsToAdd: {key: string, transformedCachedItem: TransformedCachedItem[], refType: string, from: number, to: number}[] = []
                     if(firstCharacterMatch===" " && transformedCache?.blocks?.length>0) {
                         widgetsToAdd.push({         //blocks